requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
distlib==0.3.4
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.0
filelock==3.4.2
Flask==2.0.2
greenlet==1.1.2
h11==0.13.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
orjson==3.6.7
platformdirs==2.4.1
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
virtualenv==20.12.1
virtualenvwrapper-win==1.2.7
watchgod==0.7
websockets==10.2
Werkzeug==2.0.2
